LayoutTests/imported/w3c/web-platform-tests/import-maps/multiple-import-maps/with-errors-expected.txt

CONSOLE MESSAGE: ImportMap has invalid JSON

PASS Second import map should be rejected after an import map with errors
